Setup sign in / up modals to use parallel routing
