+++ frontend/src/app/@modal/default.tsx
@@ -1,0 +1,1 @@
+export default function () { return null }
