+++ frontend/src/app/page.client.tsx
@@ -41,7 +41,8 @@ export function TagFilterList ({group, showCounts, activeTags, startChecked}: Re
             <div className={'collapse-content flex flex-wrap gap-6 justify-around'}>
                 {group.tags
                     .sort((a: Tag, b: Tag) => b.count - a.count)
-                    .map((tag: Tag) => <TagButton tag={tag} key={tag.tagId}
+                    .map((tag: Tag) => <TagButton tag={tag}
+                                                  key={tag.tagId}
                                                   checked={activeTags?.has(tag.tagId)}
                                                   showCount={showCounts}
                                                   handleChanged={handleTagButtonChanged}/>)
